Add Makefile command to run the app locally
It is no longer possible to run the app locally with
`bundle exec middleman server` because changes made in
https://github.com/alphagov/notifications-tech-docs/pull/7 mean
that `tech-docs.yml` either won't exist or may not be up to date.

The new Makefile command to run the app ensures that `tech-docs.yml` is
created and then runs `bundle exec middleman server`.