Decouple cf-deploy target from generating build files
The point of using docker is so that we don't have to install ruby and
its dependencies on our machine.

This commit removes the link between the `cf-deploy` target
and the `generate-build-files` one and instead checks if the build
directory exists before the deployment and removes it in the end to
ensure that we always have a fresh copy of the data.