Add support for multiple build directories
Rename `build/` directory to `build-CF_SPACE/` which will allow to
generate separate outputs for the different environments.

During the master build, we will generate the output for both preview
and production in different directories and add them to the zip file to
be deployed.

When deploying we copy the `build-CF_SPACE/` to `build/`, run the
deployment and then remove it again.