Pin gems in the Gemfile more precisely
We had the `govuk_tech_docs` gem in the Gemfile as
`gem 'govuk_tech_docs', '~> 1.3.1'`. This syntax allowed the minor
version to be increased automatically to version 1.6.0 when the app
was deployed, which caused a breaking change. This commit changes the
Gemfile to pin the minor versions of all the gems to avoid the risk of
any future breaking changes.