Update link to Slack channel
Updates the Slack links in the footer and on the Support page to use the Notify Slack channel's unique ID.

This is needed because the current link uses the channel name, which isn't recognised. Instead of opening the Notify Slack channel, the link opens the GDS slack and defaults to whatever channel the user last had open.