Remove local development docker support
This was previously used for jenkins to run and build the tech docs in
containers to isolate it's dependancies. This now all takes places using
our concourse pipeline so we no longer need to run these commands as
part of any CI process.

These commands could still be used for local development however I don't
think any developers are using them for that. Furthermore, there are no
tricky OS level dependancies to install so I don't think developers have
too much issue running this without docker locally. The only exception
to this could be is matching the Ruby version to the `.ruby_version`
file.

Overall, I think it's OK to remove this code as it's not providing
enough benefit as is.