Add column heading markup to tables
Adding `#` before the content inside the first cell of each row in
a table changes the html from `<td>` to `<th scope="row">` when
using version 2.2.0 or later of the tech docs.

We need this on some tables to make them more accessible.

The smoke testing phone numbers and email addresses don't need to be in
a table, since there is only one column so that changes them to bullet
points instead.