Standardise README section on running the app
The part about what a local development server is should be common
knowledge for the people making changes to this repo - it matches
the behaviour of all of our other repos. I've left the part about
the port in there as it's actionable info and it might be unclear
which port to use, since multiple ports are shown in the output.