Switch to local development with Docker
It seems a shame to ask a developer to setup a Ruby environment
just to run the code in this repo. We only have one other Ruby
repo, which is the basis for the approach I'm taking here [1].
Many of the people who develop on this repo should already have
a Docker environment, since it's required for lots of other repos,
so I reckon this change isn't too controversial.

[1]: https://github.com/alphagov/notifications-ruby-client/pull/122/commits/88cdfbb5e3f202e41438f6cdc203917a0d57d5fe