Remove custom `nginx.conf` file
We don't need a custom nginx.conf file because the Cloud Foundry
static buildpack comes with the file. Other tech docs don't specify
their own file, and the docs can be deployed and run without it. There
is also a warning when deploying with this file:

```
**WARNING** overriding nginx.conf is deprecated and highly discouraged,
as it breaks the functionality of the Staticfile and Staticfile.auth
configuration directives. Please use the NGINX buildpack available at:
https://github.com/cloudfoundry/nginx-buildpack
```

Removing the custom file will let us potentially add a new nginx.conf
file just for ECS (which can be added to `.cfignore` to ensure it's not
used on the PaaS too).