Add nginx.conf file for running on ECS
We are going to be running this app in ECS using NGINX in Docker. This
adds an `nginx.conf` file to use. The settings for it are a mix of the
settings in the file we currently use on the PaaS (in the static
buildback) and the default file that comes with the NGINX Docker image.

It also adds the same security headers and values used by GOV.UK Pay.

We don't want to start using this file on the PaaS, so by adding to the
`.cfignore` we can ensure that we don't overwrite the default static
buildpack file.

The static files for both preview and production are available and we
decide which to serve based on the `server_name`.