Update Dockerfile to have a production build stage
When running in production, we want to use NGINX as a server instead of
Middleman. This turns the Dockerfile into a multi-stage build.

The Ruby build is used when running locally and to generate the static build
files for both preview and production. If using the `run_with_docker.sh` script
to run locally we don't need build files - we mount the local files
instead.

The NGINX build copies the static files generated from the Ruby build
and the custom `nginx.conf` file we created. NGINX is running by default
in the container, so we don't need to run a command to start it. It uses
`server_name` to decide which static files to serve.