Remove old copy about manual contact for inbound sms
We may instead want to replace this with similar content to
https://github.com/alphagov/notifications-python-client/pull/229/files
but given the headings would be slightly different and so this
isn't just a copy and paste job, I went for the quicker option.

Happy for someone to put some commits on top to change the
content if they want to, but at least this removes the incorrect
information I guess?