Add a setup section to the REST API documentation
All of our clients have a ‘Setting up the client’ section which includes
subsections about installing dependencies, API keys, etc.

In the REST API documentation that information is at a level above, not
within its own section.

This commit makes the REST API documentation consistent with the other
clients, which should make it easier to navigate.