Reformat "Send a text message" error table for readability.
For this and all following tables, I have to do table column names
in the Python file instead of, like I'd prefer to, in the partial.
This is due to a bug in Middleman that means of all the table is
in the partial, and there is a paragraph with bullet points underneath,
like we have, everything below that partial is not seen as Markdown
and not rendered properly.