Correct the error message for TooManyRequestsError.
While at the same time signifying that international_sms is one
of the possible limits.

This had to be done in so many places, thanks be to the find
and replace tool. Luckily soon, when we bring in the new
de-duplicated error tables to the rest of the docs, similar
changes will only have to be made once, yay!